+++ R/dist_discrete.R
@@ -108,7 +108,7 @@ DiscreteDistribution <- distribution_class(
       probs <- tf$reshape(probs, list(-1L, k))
 
       denss <- tf$stack(lapply(seq_len(k), function(i) {
-        tf$where(x == keras::k_constant(i), probs[, i], K$zero)
+        tf$where(x == keras3::as_tensor(i, keras3::config_floatx()), probs[, i], K$zero)
       }), axis = 1L)
 
       dens <- tf$reduce_sum(denss, axis = 1L)
@@ -136,12 +136,12 @@ DiscreteDistribution <- distribution_class(
     }
   },
   tf_make_constants = function(with_params = list()) {
-    check_installed("keras")
+    check_installed("keras3")
     params <- private$.make_params(with_params, 1)
     out <- list()
     if (length(params$probs) && !is.null(params$probs[[1L]])) {
       probs <- as.numeric(params$probs)
-      out$probs <- keras::k_constant(probs / sum(probs), shape = c(1L, length(probs)))
+      out$probs <- keras3::as_tensor(probs / sum(probs), keras3::config_floatx(), shape = c(1L, length(probs)))
     }
     out
   },
@@ -150,9 +150,10 @@ DiscreteDistribution <- distribution_class(
     k <- length(ph$probs)
     if (length(ph$probs)) {
       out <- list(
-        probs = keras::layer_dense(
+        probs = keras3::layer_dense(
           input, units = k, activation = "softmax",
-          name = paste0(name_prefix, "probs")
+          name = paste0(name_prefix, "probs"),
+          dtype = keras3::config_floatx()
         )
       )
     } else {